refactor(lottoBundle) : ProfitController에 존재하던 논리 로직 이동
