+++ src/main/java/lotto/model/LottoBundle.java
@@ -1,5 +1,7 @@
 package lotto.model;
 
+import static lotto.util.Constants.ZERO;
+
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -8,6 +10,7 @@ import java.util.Optional;
 import lotto.util.enums.LottoResult;
 
 public class LottoBundle {
+    private static final Long LONG_ZERO = 0L;
     private static final int DEFAULT_VALUE = 0;
     private static final int INCREMENT_VALUE = 1;
 
@@ -36,6 +39,23 @@ public class LottoBundle {
             updateWinningResult(point, winningResult);
         }
         return winningResult;
+    }
+
+    public long totalPrize(final Map<String, Integer> compareResult) {
+        return compareResult.entrySet().stream()
+                .filter(this::shouldIncludeEntry)
+                .mapToLong(this::calculatePrizeForEntry)
+                .sum();
+    }
+
+    private boolean shouldIncludeEntry(Map.Entry<String, Integer> entry) {
+        Optional<LottoResult> result = LottoResult.fromDescription(entry.getKey());
+        return result.isPresent() && entry.getValue() > ZERO;
+    }
+
+    private long calculatePrizeForEntry(Map.Entry<String, Integer> entry) {
+        Optional<LottoResult> result = LottoResult.fromDescription(entry.getKey());
+        return result.map(lottoResult -> (long) lottoResult.getPrize() * entry.getValue()).orElse(LONG_ZERO);
     }
 
     private Map<String, Integer> initResult() {
